Fix keyboard navigation in screen chat
